requirements.txt
torch>=2.0.0
transformers>=4.40.0
peft>=0.8.0
trl>=0.7.0
datasets>=2.16.0
bitsandbytes>=0.42.0
accelerate>=0.25.0
pyyaml>=6.0
